make clippy happy

+++ esp-idf-svc/src/espnow.rs
@@ -44,7 +44,7 @@ impl EspNow {
         let mut taken = TAKEN.lock();
 
         if *taken {
-            esp!(ESP_ERR_INVALID_STATE as i32)?;
+            esp!(ESP_ERR_INVALID_STATE)?;
         }
 
         // disable modem sleep, otherwise messages queue up and we're not able
