+++ esp-idf-svc/src/sntp.rs
@@ -129,7 +129,7 @@ impl EspSntp {
         let mut taken = TAKEN.lock();
 
         if *taken {
-            esp!(ESP_ERR_INVALID_STATE as i32)?;
+            esp!(ESP_ERR_INVALID_STATE)?;
         }
 
         let sntp = Self::init(conf)?;
